more robust file and tpm cleanup
ucred is (apparently) not available on some kernels and causes a crash
of the tao admin server. Add checks for nil ucred as a temporary
stopgap -- we need to find a better way to check caller credentials.

Also add more robust failure path in tao admin server to clean up open
files and free tpm resources. The TPM on some machines will run out of
key slots, which (apparently) can only be fixed by clearing and
resetting the tpm in the bios.

Also make ExtractPCRs public.